backend/requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=gpu
absl-py=0.13.0=py37haa95532_0
aiohttp=3.7.4.post0=py37h2bbff1b_2
aioredis=1.3.1=pypi_0
asgiref=3.4.1=pypi_0
astor=0.8.1=py37haa95532_0
astunparse=1.6.3=py_0
async-timeout=3.0.1=py37haa95532_0
attrs=21.2.0=pyhd3eb1b0_0
autobahn=21.3.1=pypi_0
automat=20.2.0=pypi_0
blas=1.0=mkl
blinker=1.4=py37haa95532_0
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py37h2bbff1b_1003
ca-certificates=2021.7.5=haa95532_1
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.5.30=py37haa95532_0
cffi=1.14.6=py37h2bbff1b_0
channels=3.0.4=pypi_0
channels-redis=3.3.0=pypi_0
chardet=4.0.0=py37haa95532_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.1=pyhd3eb1b0_0
constantly=15.1.0=pypi_0
coreapi=2.3.3=pypi_0
coreschema=0.0.4=pypi_0
coverage=5.5=py37h2bbff1b_2
cryptography=3.4.7=py37h71e12ea_0
cycler=0.10.0=py37_0
cython=0.29.24=py37hd77b12b_0
daphne=3.0.2=pypi_0
dataclasses=0.8=pyh6d0b6a4_7
django=3.2.7=pypi_0
django-cors-headers=3.8.0=pypi_0
djangorestframework=3.12.4=pypi_0
drf-yasg=1.20.0=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
gast=0.3.3=py_0
git=2.32.0=haa95532_1
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpcio=1.36.1=py37hc60d5dd_1
h5py=2.10.0=py37h5e291fa_0
hdf5=1.10.4=h7ebc959_0
hiredis=2.0.0=pypi_0
hyperlink=21.0.0=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=3.2=pyhd3eb1b0_0
importlib-metadata=4.8.1=py37haa95532_0
incremental=21.3.0=pypi_0
inflection=0.5.1=pypi_0
intel-openmp=2021.3.0=haa95532_3372
itypes=1.2.0=pypi_0
jinja2=3.0.1=pypi_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9d=h2bbff1b_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.1=py37hd77b12b_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.17.2=h23ce68f_1
libtiff=4.2.0=hd0e1b90_0
lz4-c=1.9.3=h2bbff1b_1
markdown=3.3.4=py37haa95532_0
markupsafe=2.0.1=pypi_0
matplotlib=3.4.2=py37haa95532_0
matplotlib-base=3.4.2=py37h49ac443_0
mkl=2021.3.0=haa95532_524
mkl-service=2.4.0=py37h2bbff1b_0
mkl_fft=1.3.0=py37h277e83a_2
mkl_random=1.2.2=py37hf11a4ad_0
msgpack=1.0.2=pypi_0
multidict=5.1.0=py37h2bbff1b_2
munkres=1.1.4=py_0
mysqlclient=2.0.3=pypi_0
numpy=1.20.3=py37ha4e8547_0
numpy-base=1.20.3=py37hc2deb75_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=py37_0
openssl=1.1.1l=h2bbff1b_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=21.0=pypi_0
pillow=8.3.1=py37h4fa10fc_0
pip=21.0.1=py37haa95532_0
protobuf=3.17.2=py37hd77b12b_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pyjwt=2.1.0=py37haa95532_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h6538335_2
pyreadline=2.1=py37_1
pysocks=1.7.1=py37_1
python=3.7.11=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2021.1=pypi_0
qt=5.9.7=vc14h73c81de_0
requests=2.26.0=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
ruamel-yaml=0.17.16=pypi_0
ruamel-yaml-clib=0.2.6=pypi_0
scikit-learn=0.24.2=py37hf11a4ad_1
scipy=1.7.1=py37hbe87c03_2
service-identity=21.1.0=pypi_0
setuptools=58.0.4=py37haa95532_0
sip=4.19.8=py37h6538335_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=h2bbff1b_0
sqlparse=0.4.2=pypi_0
tensorboard=2.4.0=pyhc547734_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.3.0=mkl_py37h10aaca4_0
tensorflow-base=2.3.0=eigen_py37h17acbac_0
tensorflow-estimator=2.6.0=pyh7b7c402_0
tensorflow-gpu=2.3.0=he13fc11_0
termcolor=1.1.0=py37haa95532_1
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h2bbff1b_0
tornado=6.1=py37h2bbff1b_0
tqdm=4.62.2=pyhd3eb1b0_1
twisted=21.7.0=pypi_0
twisted-iocpsupport=1.0.2=pypi_0
txaio=21.2.1=pypi_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
uritemplate=3.0.1=pypi_0
urllib3=1.26.6=pyhd3eb1b0_1
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=2.0.1=pyhd3eb1b0_0
wheel=0.37.0=pyhd3eb1b0_1
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37_0
wrapt=1.12.1=py37he774522_1
xz=5.2.5=h62dcd97_0
yarl=1.6.3=py37h2bbff1b_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zope-interface=5.4.0=pypi_0
zstd=1.4.9=h19a0ad4_0
